+++ examples/src/CustomBunnymarkExample/CustomBunnymark.js
@@ -3,6 +3,7 @@ import PropTypes from "prop-types";
 import { ParticleContainer, Sprite, Text } from "react-pixi-fiber";
 import Stats from "stats.js/src/Stats";
 import * as PIXI from "pixi.js";
+import Particle from "./Particle";
 import bunnys from "./bunnys.png";
 
 const maxSize = 200000;
@@ -26,40 +27,34 @@ const generateBunny = texture => ({
   speedX: Math.random() * 10,
   speedY: Math.random() * 10 - 5,
   texture: texture,
-  x: 0,
-  y: 0,
 });
 
-const moveBunny = bunny => {
-  const movedBunny = { ...bunny };
+const moveBunny = function() {
+  this.x += this.speedX;
+  this.y += this.speedY;
+  this.speedY += gravity;
 
-  movedBunny.x += movedBunny.speedX;
-  movedBunny.y += movedBunny.speedY;
-  movedBunny.speedY += gravity;
-
-  if (movedBunny.x > maxX) {
-    movedBunny.speedX *= -1;
-    movedBunny.x = maxX;
-  } else if (movedBunny.x < minX) {
-    movedBunny.speedX *= -1;
-    movedBunny.x = minX;
+  if (this.x > maxX) {
+    this.speedX *= -1;
+    this.x = maxX;
+  } else if (this.x < minX) {
+    this.speedX *= -1;
+    this.x = minX;
   }
 
-  if (movedBunny.y > maxY) {
-    movedBunny.speedY *= -0.85;
-    movedBunny.y = maxY;
+  if (this.y > maxY) {
+    this.speedY *= -0.85;
+    this.y = maxY;
     if (Math.random() > 0.5) {
-      movedBunny.speedY -= Math.random() * 6;
+      this.speedY -= Math.random() * 6;
     }
-  } else if (movedBunny.y < minY) {
-    movedBunny.speedY = 0;
-    movedBunny.y = minY;
+  } else if (this.y < minY) {
+    this.speedY = 0;
+    this.y = minY;
   }
-
-  return movedBunny;
 };
 
-class Bunnymark extends Component {
+class CustomBunnymark extends Component {
   state = {
     bunnys: [],
     currentTexture: 0,
@@ -93,21 +88,23 @@ class Bunnymark extends Component {
 
   animate = () => {
     const { bunnys, currentTexture, isAdding } = this.state;
-    const addedBunnys = [];
 
     this.stats.update();
 
     if (isAdding) {
+      const addedBunnys = [];
+
       if (bunnys.length < maxSize) {
         for (let i = 0; i < bunniesAddedPerFrame; i++) {
           addedBunnys.push(generateBunny(currentTexture));
         }
       }
+      const newBunnys = bunnys.concat(addedBunnys);
+
+      this.setState({ bunnys: newBunnys });
     }
 
-    const newBunnys = bunnys.concat(addedBunnys).map(moveBunny);
-
-    this.setState({ bunnys: newBunnys });
+    this.particleContainer.children.forEach(bunny => bunny.update(bunny));
   };
 
   handlePointerDown = () => {
@@ -127,9 +124,20 @@ class Bunnymark extends Component {
 
     return (
       <Fragment>
-        <ParticleContainer maxSize={maxSize} properties={particleContainerProperties}>
+        <ParticleContainer
+          ref={c => (this.particleContainer = c)}
+          maxSize={maxSize}
+          properties={particleContainerProperties}
+        >
           {bunnys.map((bunny, i) => (
-            <Sprite key={i} anchor={bunnyAnchor} texture={this.bunnyTextures[bunny.texture]} x={bunny.x} y={bunny.y} />
+            <Particle
+              key={i}
+              anchor={bunnyAnchor}
+              update={moveBunny}
+              speedX={bunny.speedX}
+              speedY={bunny.speedY}
+              texture={this.bunnyTextures[bunny.texture]}
+            />
           ))}
         </ParticleContainer>
         <Text text={`${bunnys.length} BUNNIES`} style={{ fill: 0xffff00, fontSize: 14 }} x={5} y={5} />
@@ -147,8 +155,8 @@ class Bunnymark extends Component {
     );
   }
 }
-Bunnymark.contextTypes = {
+CustomBunnymark.contextTypes = {
   app: PropTypes.object,
 };
 
-export default Bunnymark;
+export default CustomBunnymark;