+++ examples/src/CustomBunnymarkExample/index.js
@@ -1,19 +1,19 @@
 import React, { Component } from "react";
 import { Stage } from "react-pixi-fiber";
-import Bunnymark from "./Bunnymark";
+import CustomBunnymark from "./CustomBunnymark";
 
 const OPTIONS = {
   backgroundColor: 0x1099bb,
 };
 
-class BunnymarkExample extends Component {
+class CustomBunnymarkExample extends Component {
   render() {
     return (
       <Stage width={800} height={600} options={OPTIONS}>
-        <Bunnymark />
+        <CustomBunnymark />
       </Stage>
     );
   }
 }
 
-export default BunnymarkExample;
+export default CustomBunnymarkExample;